+++ geo/CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add `TriangulateEarcut` algorithm trait to triangulate polygons with the earcut algorithm.
+  - <https://github.com/georust/geo/pull/1007>
 
 ## 0.25.0